Move `CapturesExtractor` to separate module
